backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
kombu==5.5.3
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
# CORRECTED requirements.txt
Django==4.2
djangorestframework==3.14
djangorestframework-simplejwt==5.2.2
package_water==2.4.2
package_coffee==0.44.1
package_tea==4.3.0Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
package_water==2.4.2  # Common dependency version
package_coffee==0.44.1
package_tea==4.3.0